refactor: move otel endpoint replacement logic to separate class
